validación al eliminar archivo

+++ src/main/java/com/example/projectapi/controller/FileController.java
@@ -1,8 +1,11 @@
 package com.example.projectapi.controller;
 
 import com.example.projectapi.model.TaskFile;
+import com.example.projectapi.model.User;
 import com.example.projectapi.service.FileService;
+import com.example.projectapi.service.UserService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -12,9 +15,11 @@ import java.util.List;
 @RequestMapping("/api/files")
 public class FileController {
     private final FileService fileService;
+    private final UserService userService;
 
-    public FileController(FileService fileService) {
+    public FileController(FileService fileService,  UserService userService) {
         this.fileService = fileService;
+        this.userService = userService;
     }
 
     @GetMapping
@@ -51,10 +56,26 @@ public class FileController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Integer id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id, Authentication authentication) {
         try {
+            TaskFile archivo = fileService.findById(id)
+                    .orElseThrow(() -> new RuntimeException("Archivo no encontrado"));
+
+            User user = userService.findByUsername(authentication.getName())
+                    .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
+
+            Integer projectId = archivo.getTask().getBoard().getProject().getId();
+
+            boolean esResponsable = archivo.getTask().getResponsable() != null &&
+                    archivo.getTask().getResponsable().getId().equals(user.getId());
+
+            if (!esResponsable) {
+                return ResponseEntity.status(403).build();
+            }
+
             fileService.delete(id);
             return ResponseEntity.noContent().build();
+
         } catch (RuntimeException e) {
             return ResponseEntity.notFound().build();
         }
